#63 size_t dosn't work in this case, threadarc.h. With type uint work properly.

+++ exchange/threadarc.h
@@ -49,9 +49,9 @@ private:
     bool abort;
     bool restart;
 
-    size_t max_tryCounter;
+    uint max_tryCounter;
 
-    size_t mdelay;// milliseconds delay
+    uint mdelay;// milliseconds delay
 
     Status_t status;
     QVarLengthArray<ComDataReq_t,1024> array;
